move tile device from package to base
