+++ bitblas/base/roller/__init__.py
@@ -4,4 +4,4 @@ from .node import PrimFuncNode  # noqa: F401
 from .rasterization import NoRasterization, Rasterization2DRow, Rasterization2DColumn  # noqa: F401
 from .hint import Hint  # noqa: F401
 from .policy import DefaultPolicy, TensorCorePolicy  # noqa: F401
-from .arch import TileDevice, CUDA  # noqa: F401
+from ..arch import TileDevice, CUDA  # noqa: F401
